+++ index.d.ts
@@ -7,6 +7,7 @@ export * from "./types/Callout";
 export * from "./types/Checkbox";
 export * from "./types/ColorPicker";
 export * from "./types/DatePicker";
+export * from "./types/DateTimePicker";
 export * from "./types/Dropdown";
 export * from "./types/MultiEmailInput";
 export * from "./types/Input";
@@ -19,6 +20,7 @@ export * from "./types/Pane";
 export * from "./types/Popover";
 export * from "./types/Radio";
 export * from "./types/Select";
+export * from "./types/Slider";
 export * from "./types/Spinner";
 export * from "./types/Stepper";
 export * from "./types/Switch";
@@ -27,11 +29,9 @@ export * from "./types/Table";
 export * from "./types/Tag";
 export * from "./types/Textarea";
 export * from "./types/TimePicker";
+export * from "./types/TimePickerInput";
 export * from "./types/Toastr";
 export * from "./types/Tooltip";
 export * from "./types/Tree";
 export * from "./types/TreeSelect";
 export * from "./types/Typography";
-export * from "./types/Slider";
-export * from "./types/TimePickerInput";
-export * from "./types/DateTimePicker";
