+++ package.json
@@ -15,6 +15,7 @@
   },
   "files": [
     "dist",
+    "types",
     "*.d.ts"
   ],
   "lint-staged": {
@@ -40,8 +41,8 @@
       "types": "./layouts.d.ts"
     },
     "./formik": {
-      "import": "./dist/formik.js",
-      "require": "./dist/cjs/formik.js",
+      "import": "./dist/formik/index.js",
+      "require": "./dist/cjs/formik/index.js",
       "types": "./formik.d.ts"
     },
     "./managers": {
@@ -52,12 +53,12 @@
     "./*": {
       "import": "./dist/*.js",
       "require": "./dist/cjs/*.js",
-      "types": "./index.d.ts"
+      "types": "./types/*.d.ts"
     },
     "./formik/*": {
       "import": "./dist/formik/*.js",
       "require": "./dist/cjs/formik/*.js",
-      "types": "./formik.d.ts"
+      "types": "./types/formik/*.d.ts"
     }
   },
   "devDependencies": {
