+++ src/components/index.js
@@ -48,6 +48,7 @@ export {
   Checkbox,
   ColorPicker,
   DatePicker,
+  DateTimePicker,
   Dropdown,
   MultiEmailInput,
   Input,
@@ -60,6 +61,7 @@ export {
   Popover,
   Radio,
   Select,
+  Slider,
   Spinner,
   Stepper,
   Switch,
@@ -68,12 +70,10 @@ export {
   Tag,
   Textarea,
   TimePicker,
+  TimePickerInput,
   Toastr,
   Tooltip,
   Tree,
   TreeSelect,
   Typography,
-  Slider,
-  TimePickerInput,
-  DateTimePicker,
 };
