+++ types/formik/ActionBlock.d.ts
@@ -1,0 +1,10 @@
+import React from "react";
+import { ButtonProps } from "../Button";
+
+export interface ActionBlockProps {
+  className?: string;
+  submitButtonProps?: ButtonProps;
+  cancelButtonProps?: ButtonProps;
+  isSubmitting?: boolean;
+}
+export const ActionBlock: React.FC<ActionBlockProps>;
