+++ types/formik/Button.d.ts
@@ -1,0 +1,3 @@
+import { Button as PlainButton } from "../Button";
+
+export const Button: typeof PlainButton;
