+++ types/formik/Checkbox.d.ts
@@ -1,0 +1,3 @@
+import { Checkbox as PlainCheckbox } from "../Checkbox";
+
+export const Checkbox: typeof PlainCheckbox;
