+++ types/formik/Input.d.ts
@@ -1,0 +1,3 @@
+import { Input as PlainInput } from "../Input";
+
+export const Input: typeof PlainInput;
