+++ types/formik/MultiEmailInput.d.ts
@@ -1,0 +1,3 @@
+import { MultiEmailInput as PlainMultiEmailInput } from "../MultiEmailInput";
+
+export const MultiEmailInput: typeof PlainMultiEmailInput;
