+++ types/formik/Radio.d.ts
@@ -1,0 +1,3 @@
+import { Radio as PlainRadio } from "../Radio";
+
+export const Radio: typeof PlainRadio;
