+++ types/formik/Select.d.ts
@@ -1,0 +1,3 @@
+import { Select as PlainSelect } from "../Select";
+
+export const Select: typeof PlainSelect;
