+++ types/formik/Slider.d.ts
@@ -1,0 +1,3 @@
+import { Slider as PlainSlider } from "../Slider";
+
+export const Slider: typeof PlainSlider;
