+++ types/formik/Switch.d.ts
@@ -1,0 +1,3 @@
+import { Switch as PlainSwitch } from "../Switch";
+
+export const Switch: typeof PlainSwitch;
