+++ types/formik/Textarea.d.ts
@@ -1,0 +1,3 @@
+import { Textarea as PlainTextarea } from "../Textarea";
+
+export const Textarea: typeof PlainTextarea;
