+++ types/formik/TreeSelect.d.ts
@@ -1,0 +1,3 @@
+import { TreeSelect as PlainTreeSelect } from "../TreeSelect";
+
+export const TreeSelect: typeof PlainTreeSelect;
